LayoutTests/media/modern-media-controls/macos-media-controls/macos-media-controls-expected.txt

Testing the MacOSMediaControls constructor.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS mediaControls.element.localName is "div"
PASS mediaControls.element.className is "media-controls mac"
PASS mediaControls.layoutTraits is LayoutTraits.macOS
PASS mediaControls.muteButton instanceof MuteButton is true
PASS mediaControls.tracksButton instanceof TracksButton is true
PASS mediaControls.volumeSlider instanceof VolumeSlider is true

PASS mediaControlsWithArguments.width is 600
PASS mediaControlsWithArguments.height is 300

PASS successfullyParsed is true

TEST COMPLETE
